Implemented rotating of objects. It works only in edit mode at the moment. For presentation mode it follows soon. Rotating of cliparts doesn't work well at the moment - I'll work on that.
svn path=/trunk/koffice/; revision=9828

+++ kpresenter/qwmf.cc
@@ -165,7 +165,7 @@ bool QWinMetaFile::load(const QString aFileName)
 
     if (mSingleStep)
     {
-      debug("WMF Placeable Header (%d):", sizeof(pheader));
+      debug("WMF Placeable Header (%d):", (int)sizeof(pheader));
       debug("  key=%x", pheader.key);
       debug("  hmf=%x", pheader.hmf);
       debug("  bbox=(%d;%d;%d;%d)", pheader.bbox.left, pheader.bbox.top,
